SRMAP-Hybrid-CaptchaSolver/requirements.txt
torch>=1.13
torchvision>=0.14
fastapi>=0.95
uvicorn>=0.22
pillow>=9.0
pandas>=1.5
python-multipart>=0.0.6
onnxruntime>=1.16